requirements.txt
click>=8.0.0
requests>=2.28.0
python-dotenv>=0.19.0
websockets>=10.0.0
loguru>=0.6.0